+++ Newspack/Newspack/View/MenuTableViewCell.swift
@@ -1,0 +1,6 @@
+import Foundation
+import UIKit
+
+class MenuTableViewCell: UITableViewCell {
+
+}
